feat: Add Client and Meter routes

+++ src/api/user/user.controller.ts
@@ -7,6 +7,7 @@ import {
   handleDeleteUser,
   handleUserLogin,
 } from "./user.service";
+import { AuthenticatedRequest } from "../../types/AuthenticatedRequest";
 
 export const getAllUsers = async (req: Request, res: Response) => {
   const data = await handleGetAllUsers();
@@ -18,20 +19,36 @@ export const createUser = async (req: Request, res: Response) => {
   res.status(data.success ? 200 : 400).send(data);
 };
 
-export const getUserById = async (req: Request, res: Response) => {
-  const { userId } = req.params;
+export const getUserById = async (req: AuthenticatedRequest, res: Response) => {
+  const userId = req.user?.id;
+
+  if (!userId) {
+    res.status(401).json({ success: false, message: "User not authenticated" });
+    return; // Use `return` to exit the function
+  }
+
   const data = await handleGetUserById(userId);
   res.status(data.success ? 200 : 404).send(data);
 };
 
-export const updateUser = async (req: Request, res: Response) => {
-  const { userId } = req.params;
+export const updateUser = async (req: AuthenticatedRequest, res: Response) => {
+  const userId = req.user?.id;
+
+  if (!userId) {
+    res.status(401).json({ success: false, message: "User not authenticated" });
+    return; // Use `return` to exit the function
+  }
   const data = await handleUpdateUser(userId, req.body);
   res.status(data.success ? 200 : 400).send(data);
 };
 
-export const deleteUser = async (req: Request, res: Response) => {
-  const { userId } = req.params;
+export const deleteUser = async (req: AuthenticatedRequest, res: Response) => {
+  const userId = req.user?.id;
+
+  if (!userId) {
+    res.status(401).json({ success: false, message: "User not authenticated" });
+    return; // Use `return` to exit the function
+  }
   const data = await handleDeleteUser(userId);
   res.status(data.success ? 200 : 400).send(data);
 };
